nav bar made consistent. Corner cases added in student view

+++ pisces/students/views.py
@@ -78,7 +78,7 @@ class logout(TemplateView):
             return HttpResponse(status=400)
 
     def post(self, request):
-        return HttpResponse(status=400)
+        return HttpResponse(status=403)
 
 
 class updatePassword(TemplateView):
@@ -121,50 +121,45 @@ class Company_list(TemplateView):
     template_name = 'company.html'
 
     def get(self, request):
-        t = datetime.date(datetime.now())
-        # t = t.strftime("%Y-%m-%d")
-        # t = datetime.strptime(t,"%Y-%m-%d")
-        print(type(t))
-        # companies = Company.objects.filter(__raw__={"Registration Deadline":{"lte":t}})
-        companies = Company.objects.filter()
-        # print(companies)
-        print('This is the registration Deadline', type(companies[1]['Registration Deadline']),
-              companies[1]['Registration Deadline'])
+        if('username' in request.session):
+            t = datetime.date(datetime.now())
 
-        details = json.loads(companies.to_json())
-        df = pd.DataFrame(details)
-        # print(df.columns)
+            print(type(t))
+            companies = Company.objects.filter()
+            print('This is the registration Deadline', type(companies[1]['Registration Deadline']),
+                  companies[1]['Registration Deadline'])
 
-        # print(details)
-        # df = df.dropna(subset=['Registration Deadline'])
+            details = json.loads(companies.to_json())
+            df = pd.DataFrame(details)
 
-        df = df[df['Registration Deadline'] != '']
-        deadlines = list(df['Registration Deadline'])
-        deadlines = [i.split('T')[0] for i in deadlines]
-        deadlines = [datetime.strptime(i, "%Y-%m-%d") for i in deadlines]
-        df['Registration Deadline'] = deadlines
+            df = df[df['Registration Deadline'] != '']
+            deadlines = list(df['Registration Deadline'])
+            deadlines = [i.split('T')[0] for i in deadlines]
+            deadlines = [datetime.strptime(i, "%Y-%m-%d") for i in deadlines]
+            df['Registration Deadline'] = deadlines
 
-        df['Registration Deadline'] = pd.to_datetime(df['Registration Deadline'], format='%Y-%m-%d')
-        print(df['Registration Deadline'][1], pd.Timestamp(t), df['Registration Deadline'][1] > pd.Timestamp(t))
-        # df = df[df['Registration Deadline'] >= pd.Timestamp(t)]
-        df = df.sort_values(by="Registration Deadline",ascending = False)
-        details = list(df.T.to_dict().values())
-        flag = 0
-        flag2 = 0
-        for i in range(len(details)):
-            details[i]['Compensation'] = dict(details[i]['Compensation'])
+            df['Registration Deadline'] = pd.to_datetime(df['Registration Deadline'], format='%Y-%m-%d')
+            print(df['Registration Deadline'][1], pd.Timestamp(t), df['Registration Deadline'][1] > pd.Timestamp(t))
+            df = df.sort_values(by="Registration Deadline",ascending = False)
+            details = list(df.T.to_dict().values())
+            flag = 0
+            flag2 = 0
+            for i in range(len(details)):
+                details[i]['Compensation'] = dict(details[i]['Compensation'])
 
-            # print(details[i]['_id'],type(details[i]['_id']))
-            details[i]['id'] = details[i]['_id']['$oid']
-            if details[i]['Registration Deadline'] >= pd.Timestamp(t):
-                details[i]['allowed']=1
-            else:
-                if(flag ==0):
-                    flag=1
-                    details[i]['first'] = 1
-                details[i]['allowed']=0
-            # print(type(details[i]['Compensation']))
-        return render(request, self.template_name, {'details': details, 'notRendered': ["_id", "id", ]})
+                # print(details[i]['_id'],type(details[i]['_id']))
+                details[i]['id'] = details[i]['_id']['$oid']
+                if details[i]['Registration Deadline'] >= pd.Timestamp(t):
+                    details[i]['allowed']=1
+                else:
+                    if(flag ==0):
+                        flag=1
+                        details[i]['first'] = 1
+                    details[i]['allowed']=0
+                # print(type(details[i]['Compensation']))
+            return render(request, self.template_name, {'details': details, 'notRendered': ["_id", "id", ]})
+        else:
+            return redirect('/accounts/login')
 
 
 class register(TemplateView):
@@ -289,32 +284,35 @@ class register(TemplateView):
 
             else:
                 # Not meeting eligibility criteria
-                return HttpResponse(status=200)
-            return HttpResponse(status=200)
+                return render(request, 'icant.html')
+            return render(request, 'icant.html')
 
-        return HttpResponse(status=400)
+        return redirect('/accounts/login')
 
 
 class uploadResume(TemplateView):
     def post(self, request):
-        uploaded_file = request.FILES['resume']
-        print(uploaded_file.name)
-        print(uploaded_file.size)
-        fs = FileSystemStorage()
-        if ('pdf' in uploaded_file.name or 'doc' in uploaded_file.name):
-            files = set(os.listdir(settings.MEDIA_ROOT))
-            print(files)
-            for i in files:
-                if request.session['username'] in i:
-                    os.remove(settings.MEDIA_ROOT + '/' + i)
-            fileName = request.session['username'] + '.' + uploaded_file.name.split('.')[-1]
-            fs.save(fileName, uploaded_file)
-            userUpload = User.objects.get(srn=request.session['username'])
-            userUpload.update(set__file=fileName)
-            return redirect('/students/homepage')
+        if('username' in request.session):
+            uploaded_file = request.FILES['resume']
+            print(uploaded_file.name)
+            print(uploaded_file.size)
+            fs = FileSystemStorage()
+            if ('pdf' in uploaded_file.name or 'doc' in uploaded_file.name):
+                files = set(os.listdir(settings.MEDIA_ROOT))
+                print(files)
+                for i in files:
+                    if request.session['username'] in i:
+                        os.remove(settings.MEDIA_ROOT + '/' + i)
+                fileName = request.session['username'] + '.' + uploaded_file.name.split('.')[-1]
+                fs.save(fileName, uploaded_file)
+                userUpload = User.objects.get(srn=request.session['username'])
+                userUpload.update(set__file=fileName)
+                return redirect('/students/homepage')
+            else:
+                messages.info(request, 'File not uploaded')
+                return redirect('/students/homepage')
         else:
-            messages.info(request, 'File not uploaded')
-            return redirect('/students/homepage')
+            redirect('accounts/login')
 
 class viewSchedule(TemplateView):
     template_name = 'schedule.html'
@@ -369,56 +367,64 @@ class Blog(TemplateView):
     template_name = 'blog-posts.html'
 
     def get(self,request):
-        form = BlogPost()
-        blogs = blogging.objects.filter().order_by('-created_on')
+        if('username' in request.session):
+            form = BlogPost()
+            blogs = blogging.objects.filter().order_by('-created_on')
 
-        # print(blogs.to_json())
-        #print(blogs[0].updated_on)
-        blogs = json.loads(blogs.to_json())
-        blogs = blogs[::-1]
-        for i in blogs:
-            # print(i['author'])
-            i['author']=User.objects.get(id=i['author']['$oid'])['name']
-            del i['content']
-        print(len(blogs))
-        return render(request, self.template_name, {"form": form,"blogs":blogs})
+            # print(blogs.to_json())
+            #print(blogs[0].updated_on)
+            blogs = json.loads(blogs.to_json())
+            blogs = blogs[::-1]
+            for i in blogs:
+                # print(i['author'])
+                i['author']=User.objects.get(id=i['author']['$oid'])['name']
+                del i['content']
+            print(len(blogs))
+            return render(request, self.template_name, {"form": form,"blogs":blogs})
+        else:
+            return redirect('/accounts/login')
     def post(self,request):
-        print(request.POST)
-        user = User.objects.get(srn=request.session['username'])
-        content = dict()
-        values = dict(request.POST)
-        del values['csrfmiddlewaretoken']
-        del values['blogTitle']
-        del values['company']
-        del values['blogType']
-        del values['shortDescription']
-        content = dict()
-        for i in values:
-            if('type' in i):
-                count = i.split('type')[1]
-                if(count not in content):
-                    content[count] = ['','']
-                content[count][0]=values[i]
-            elif('text' in i):
-                count = i.split('text')[1]
-                if(count not in content):
-                    content[count]=['','']
-                content[count][1] = values[i]
-        print(content)
-        keys = list(content.keys())
-        keys.sort()
-        finalContent = []
-        for i in keys:
-            finalContent.append(content[i])
-        newBlog = blogging()
-        newBlog.title = request.POST['blogTitle']
-        newBlog.author = user['id']
-        newBlog.company = request.POST['company']
-        newBlog.blogType = request.POST['blogType']
-        newBlog.shortDescription = request.POST['shortDescription']
-        newBlog.content = finalContent
-        newBlog.save(force_insert=True)
-        return redirect('/students/blog')
+        if('username' in request.session):
+            print(request.POST)
+            user = User.objects.get(srn=request.session['username'])
+            content = dict()
+            values = dict(request.POST)
+            del values['csrfmiddlewaretoken']
+            del values['blogTitle']
+            del values['company']
+            del values['blogType']
+            del values['shortDescription']
+            content = dict()
+            for i in values:
+                if('type' in i):
+                    count = i.split('type')[1]
+                    if(count not in content):
+                        content[count] = ['','']
+                    content[count][0]=values[i]
+                elif('text' in i):
+                    count = i.split('text')[1]
+                    if(count not in content):
+                        content[count]=['','']
+                    content[count][1] = values[i]
+            print(content)
+            keys = list(content.keys())
+            keys.sort()
+            finalContent = []
+            for i in keys:
+                finalContent.append(content[i])
+            newBlog = blogging()
+            newBlog.title = request.POST['blogTitle']
+            newBlog.author = user['id']
+            newBlog.company = request.POST['company']
+            newBlog.blogType = request.POST['blogType']
+            newBlog.shortDescription = request.POST['shortDescription']
+            newBlog.content = finalContent
+            newBlog.save(force_insert=True)
+            return redirect('/students/blog')
+        else:
+            return redirect('/accounts/login')
+
+
 class SearchBlog(TemplateView):
     template_name = 'blog-posts.html'
 
@@ -426,31 +432,37 @@ class SearchBlog(TemplateView):
         return HttpResponse(status=403)
 
     def post(self,request):
-        form = BlogPost()
-        blogs = blogging.objects.filter(company=request.POST['company']).order_by('-created_on')
+        if('username' in request.session):
+            form = BlogPost()
+            blogs = blogging.objects.filter( Q(company__contains=request.POST['company']) |  Q(title__contains=request.POST['company']) | Q(shortDescription__contains=request.POST['company'])).order_by('-created_on')
 
-        # print(blogs.to_json())
-        # print(blogs[0].updated_on)
-        blogs = json.loads(blogs.to_json())
-        blogs = blogs[::-1]
-        for i in blogs:
-            # print(i['author'])
-            i['author'] = User.objects.get(id=i['author']['$oid'])['name']
-            del i['content']
-        print(len(blogs))
-        return render(request, self.template_name, {"form": form, "blogs": blogs})
+            # print(blogs.to_json())
+            # print(blogs[0].updated_on)
+            blogs = json.loads(blogs.to_json())
+            blogs = blogs[::-1]
+            for i in blogs:
+                # print(i['author'])
+                i['author'] = User.objects.get(id=i['author']['$oid'])['name']
+                del i['content']
+            print(len(blogs))
+            return render(request, self.template_name, {"form": form, "blogs": blogs})
+        else:
+            return redirect('/accounts/login')
 
 class BlogDetails(TemplateView):
     template_name = 'blog-detail.html'
     def post(self,request):
-        print(request.POST)
-        form = BlogPost()
-        blog = blogging.objects.get(id=request.POST['id'])
-        blog['author'] = User.objects.get(id=blog['author'])['name']
-        blog = json.loads(blog.to_json())
-        blog['content'] = list(blog['content'])
-        del blog['_id']
-        return render(request,self.template_name,{'blog':blog,'form':form})
+        if('username' in request.session):
+            print(request.POST)
+            form = BlogPost()
+            blog = blogging.objects.get(id=request.POST['id'])
+            blog['author'] = User.objects.get(id=blog['author'])['name']
+            blog = json.loads(blog.to_json())
+            blog['content'] = list(blog['content'])
+            del blog['_id']
+            return render(request,self.template_name,{'blog':blog,'form':form})
+        else:
+            return redirect('/accounts/login')
 
 
 
@@ -617,7 +629,7 @@ def wordCloud(df):
     plt.close()
     return 'data:image/png;base64,{}'.format(graph_url)
 
-'''df = pd.read_csv("detail.csv")
+df = pd.read_csv("detail.csv")
 df = df.sort_values(by='Date that the company has come')
 df = df.reset_index(drop=True)
 a = list(df['CTC'])
@@ -630,30 +642,33 @@ to_be_predicted=np.array([[second,first]])
 to_be_predicted = to_be_predicted.reshape(to_be_predicted.shape[0],to_be_predicted.shape[1],1)
 model = load_model('ctc_lstm.h5')
 value = model.predict(to_be_predicted,verbose=0)
-print('This is the value',value)'''
+print('This is the value',value)
 
 
 class ViewStatistics(TemplateView):
     template_name = 'stats.html'
     def get(self,request):
-        print(os.listdir())
-        df = pd.read_csv("detail.csv")
+        if('username' in request.session):
+            print(os.listdir())
+            df = pd.read_csv("detail.csv")
 
-        df = df.sort_values(by='Date that the company has come')
-        df = df.reset_index(drop=True)
-        a = list(df['CTC'])
-        print(len(a))
-        first=float(a[-1].replace(',',''))
-        first = first/100000
-        second = float(a[-2].replace(',',''))
-        second = second/100000
-        to_be_predicted=np.array([[second,first]])
-        to_be_predicted = to_be_predicted.reshape(to_be_predicted.shape[0],to_be_predicted.shape[1],1)
-        all_plots = []
-        all_plots += ctc(df)
-        plot2 = wordCloud(df)
-        global value
-        return render(request,self.template_name,{"all_plots":all_plots,"wordCloud": plot2,'next':str(value[0][0])})
+            df = df.sort_values(by='Date that the company has come')
+            df = df.reset_index(drop=True)
+            a = list(df['CTC'])
+            print(len(a))
+            first=float(a[-1].replace(',',''))
+            first = first/100000
+            second = float(a[-2].replace(',',''))
+            second = second/100000
+            to_be_predicted=np.array([[second,first]])
+            to_be_predicted = to_be_predicted.reshape(to_be_predicted.shape[0],to_be_predicted.shape[1],1)
+            all_plots = []
+            all_plots += ctc(df)
+            plot2 = wordCloud(df)
+            global value
+            return render(request,self.template_name,{"all_plots":all_plots,"wordCloud": plot2,'next':str(value[0][0])})
+        else:
+            return redirect('/accounts/login')
 
 
 class SearchCompany(TemplateView):
